requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-bcrypt==1.0.1
werkzeug==3.0.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0